requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.42.0
anyio==4.7.0
cachetools==5.5.1
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.3.0
colorlog==6.9.0
defusedxml==0.7.1
distro==1.9.0
fastapi==0.121.1
google-api-core==2.24.0
google-api-python-client==2.159.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.66.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
jiter==0.8.2
proto-plus==1.25.0
protobuf==5.29.3
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.4
pydantic_core==2.27.2
pyparsing==3.2.1
python-dotenv==1.1.1
pytubefix==8.12.0
requests==2.32.3
rsa==4.9
sniffio==1.3.1
sprinkles==0.4.6
SQLAlchemy==2.0.44
starlette==0.49.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.38.0
workspace==0.3.1
youtube-transcript-api==1.1.1
